requirements.txt
gunicorn==19.9.0
opencv-python==4.3.0.36
numpy==1.19.1
matplotlib==3.3.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
Pillow==7.2.0
